fix(config): repair breakage from the plugin reduction
The reduce-plugins refactor left several regressions. Fixes:

- mapleader was set after config.remap was required, so every keymap in
  remap.lua bound to `\` instead of space (<leader> resolves when a mapping
  is created, not when it is pressed). Set it first.

- mason-lspconfig was given both `opts` and a `config` function. lazy.nvim
  skips its default setup(opts) call when `config` is defined, and ours never
  called it, so ensure_installed/automatic_enable were dead and nothing ever
  called vim.lsp.enable(). No server attached; `gd` did nothing.

- mini.completion popped a buffer-word menu inside snacks picker prompts,
  covering the results list. Disable it in prompt buffers.

- lazy's update checker lost `notify = false`, so it fired notifier toasts.

- Restore kanagawa (dropped despite 801b68a) and nvim-treesitter, which
  set.lua's fold options still depend on for foldexpr/highlight/indent.

- Restore mini.snippets' friendly-snippets dependency; without it
  gen_loader.from_lang() had nothing to read and no snippet expanded.

- Restore format-on-save via conform, and route <leader>F through it so
  manual formatting matches (it used lua_ls instead of stylua before).

- Drop mini.pick/mini.extra, which shadowed snacks' <leader>xx and <leader>pk.

- Delete lua/config/lsp.lua (fully commented out; live copy is in
  lua/plugins/lsp.lua) and fix format.bash's `#/bin/bash` shebang.
