+++ init.lua
@@ -1,50 +1,1 @@
-vim.loader.enable()
-
--- Make sure to setup `mapleader` and `maplocalleader` before
--- loading lazy.nvim so that mappings are correct.
--- This is also a good place to setup other settings (vim.opt)
-vim.g.mapleader = " "
-vim.g.maplocalleader = "\\"
-
-require("config.remap")
-
--- When embedded inside VS Code (vscode-neovim), skip plugins and LSP —
--- VS Code owns those. Only keymaps from remap.lua carry over.
-if vim.g.vscode then
-  return
-end
-
-require("config.set")
-require("config.lsp")
-
--- Bootstrap lazy.nvim
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-if not (vim.uv or vim.loop).fs_stat(lazypath) then
-	local lazyrepo = "https://github.com/folke/lazy.nvim.git"
-	local out = vim.fn.system({ "git", "clone", "--filter=blob:none", "--branch=stable", lazyrepo, lazypath })
-	if vim.v.shell_error ~= 0 then
-		vim.api.nvim_echo({
-			{ "Failed to clone lazy.nvim:\n", "ErrorMsg" },
-			{ out, "WarningMsg" },
-			{ "\nPress any key to exit..." },
-		}, true, {})
-		vim.fn.getchar()
-		os.exit(1)
-	end
-end
-vim.opt.rtp:prepend(lazypath)
-
--- Setup lazy.nvim
-require("lazy").setup({
-	spec = "plugins",
-	-- Configure any other settings here. See the documentation for more details.
-	-- colorscheme that will be used when installing plugins.
-	install = { colorscheme = { "habamax" } },
-	-- automatically check for plugin updates
-	checker = { enabled = true, notify = false },
-})
-
--- Our lsp/*.lua would otherwise be overridden by nvim-lspconfig's copies of the
--- same servers. Re-applying them here restores our settings; must come after
--- lazy.setup() so `require("lspconfig.util")` resolves.
-require("config.lsp").apply_local_configs()
+require("config.lazy")
