--- lsp/lua_ls.lua
@@ -1,20 +1,0 @@
----@type vim.lsp.Config
-return {
-	cmd = { "lua-language-server" },
-	filetypes = { "lua" },
-	root_markers = {
-		".luarc.json",
-		".luarc.jsonc",
-		".luacheckrc",
-		".stylua.toml",
-		"stylua.toml",
-		"selene.toml",
-		"selene.yml",
-		".git",
-	},
-	settings = {
-		Lua = {
-			diagnostics = { globals = { "vim" } },
-		},
-	},
-}
