+++ lua/config/remap.lua
@@ -70,49 +70,3 @@ vim.keymap.set("n", "<leader>u", function()
 end, { desc = "Toggle Builtin Undotree" })
 
 vim.keymap.set("n", "<leader>re", "<cmd>restart<cr>", { desc = "Restart config :restart)" })
-
--- ── VS Code (vscode-neovim) ────────────────────────────────────────────────
--- Mirrors the picker/window keymaps so muscle memory works inside VS Code.
-if vim.g.vscode then
-  local vscode = require("vscode")
-  local map = function(key, cmd, desc)
-    vim.keymap.set("n", key, function() vscode.call(cmd) end, { desc = desc })
-  end
-
-  -- File search (telescope.ff / C-p)
-  map("<leader>ff", "workbench.action.quickOpen", "Find files")
-  map("<C-p>", "workbench.action.quickOpen", "Find files")
-
-  -- Live grep (telescope.fl / fr)
-  map("<leader>fl", "workbench.action.findInFiles", "Live grep")
-  map("<leader>fr", "workbench.action.findInFiles", "Grep")
-
-  -- Grep word under cursor (telescope.fw)
-  vim.keymap.set("n", "<leader>fw", function()
-    vscode.call("workbench.action.findInFiles", { args = { query = vim.fn.expand("<cword>") } })
-  end, { desc = "Grep word under cursor" })
-
-  -- Buffer list (telescope.fb)
-  map("<leader>fb", "workbench.action.showAllEditors", "Find buffers")
-
-  -- Window navigation (C-h/j/k/l)
-  map("<C-h>", "workbench.action.focusLeftGroup", "Focus left")
-  map("<C-j>", "workbench.action.focusBelowGroup", "Focus below")
-  map("<C-k>", "workbench.action.focusAboveGroup", "Focus above")
-  map("<C-l>", "workbench.action.focusRightGroup", "Focus right")
-
-  -- Splits (ss / sv)
-  map("ss", "workbench.action.splitEditorDown", "Split down")
-  map("sv", "workbench.action.splitEditorRight", "Split right")
-
-  -- Tabs (tab / s-tab)
-  map("<tab>", "workbench.action.nextEditor", "Next editor")
-  map("<s-tab>", "workbench.action.previousEditor", "Prev editor")
-
-  -- Format (<leader>F)
-  map("<leader>F", "editor.action.formatDocument", "Format document")
-
-  -- Quickfix nav (<leader>cn / cp)
-  map("<leader>cn", "editor.action.marker.nextInFiles", "Next problem")
-  map("<leader>cp", "editor.action.marker.prevInFiles", "Prev problem")
-end
