--- lua/plugins/blink.lua
@@ -1,31 +1,0 @@
-return {
-	"saghen/blink.cmp",
-	event = "InsertEnter",
-	dependencies = { "nvim-mini/mini.snippets" },
-	version = "1.*",
-	---@module 'blink.cmp'
-	---@type blink.cmp.Config
-	opts = {
-		keymap = {
-			preset = "none",
-			["<C-space>"] = { "show", "fallback" },
-			["<C-n>"] = { "select_next", "fallback" },
-			["<C-p>"] = { "select_prev", "fallback" },
-			["<C-y>"] = { "select_and_accept", "fallback" },
-			["<C-e>"] = { "cancel", "fallback" },
-		},
-		appearance = {
-			nerd_font_variant = "mono",
-		},
-		completion = {
-			accept = { auto_brackets = { enabled = true } },
-			documentation = { auto_show = true },
-		},
-		snippets = { preset = "mini_snippets" },
-		sources = {
-			default = { "lsp", "path", "snippets", "buffer" },
-		},
-		fuzzy = { implementation = "prefer_rust" },
-	},
-	opts_extend = { "sources.default" },
-}
