--- lua/plugins/curl.lua
@@ -1,8 +1,0 @@
-return {
-	"oysandvik94/curl.nvim",
-	cmd = { "CurlOpen" },
-	dependencies = {
-		"nvim-lua/plenary.nvim",
-	},
-	config = true,
-}
