--- lua/plugins/diagnostics.lua
@@ -1,16 +1,0 @@
-return {
-	{
-		"rachartier/tiny-inline-diagnostic.nvim",
-		event = "VeryLazy",
-		opts = {
-			preset = "modern",
-			options = {
-				-- show the diagnostic inline on every line, not only the cursor line
-				multilines = {
-					enabled = true,
-					always_show = true,
-				},
-			},
-		},
-	},
-}
