--- lua/plugins/direnv.lua
@@ -1,4 +1,0 @@
-return {
-	"direnv/direnv.vim",
-	event = "VeryLazy",
-}
