--- lua/plugins/lualine.lua
@@ -1,21 +1,0 @@
-return {
-	"nvim-lualine/lualine.nvim",
-	dependencies = { "nvim-tree/nvim-web-devicons" },
-	event = "VeryLazy",
-	opts = {
-		options = {
-			theme = "auto",
-			component_separators = { left = "", right = "" },
-			section_separators = { left = "", right = "" },
-			globalstatus = true,
-		},
-		sections = {
-			lualine_a = { "mode" },
-			lualine_b = { "branch", "diff", "diagnostics" },
-			lualine_c = { { "filename", path = 1 } },
-			lualine_x = { "encoding", "filetype" },
-			lualine_y = { "progress" },
-			lualine_z = { "location" },
-		},
-	},
-}
